# backend/swagger/swagger.json
{
    "openapi": "3.0.0",
    "paths": {
        "/cars": {
            "get": {
                "operationId": "findAll",
                "parameters": [
                    {
                        "name": "q",
                        "required": false,
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            },
            "post": {
                "operationId": "createOne",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/CreateCarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            },
            "put": {
                "operationId": "updateOne",
                "parameters": [],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateCarRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/BadRequestResponse"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            }
        },
        "/cars/{id}": {
            "get": {
                "operationId": "findOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            },
            "delete": {
                "operationId": "deleteOne",
                "parameters": [
                    {
                        "name": "id",
                        "required": true,
                        "in": "path",
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "deleted",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NotFoundResponse"
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Cars"
                ]
            }
        },
        "/models": {
            "get": {
                "operationId": "findAll",
                "parameters": [],
                "responses": {
                    "200": {
                        "description": "",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Model"
                                    }
                                }
                            }
                        }
                    }
                },
                "tags": [
                    "Models"
                ]
            }
        }
    },
    "info": {
        "title": "Cars Project",
        "description": "The Cars Project API description",
        "version": "1.0",
        "contact": {}
    },
    "tags": [
        {
            "name": "Cars",
            "description": ""
        },
        {
            "name": "Models",
            "description": ""
        }
    ],
    "servers": [],
    "components": {
        "schemas": {
            "Model": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name"
                ]
            },
            "Car": {
                "type": "object",
                "properties": {
                    "id": {
                        "type": "number"
                    },
                    "name": {
                        "type": "string"
                    },
                    "model": {
                        "$ref": "#/components/schemas/Model"
                    },
                    "color": {
                        "type": "string"
                    },
                    "year": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string"
                    }
                },
                "required": [
                    "id",
                    "name",
                    "model",
                    "color",
                    "year",
                    "image"
                ]
            },
            "NotFoundResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number",
                        "default": 404
                    },
                    "message": {
                        "type": "string",
                        "default": "Not Found"
                    }
                },
                "required": [
                    "statusCode",
                    "message"
                ]
            },
            "CreateCarRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "modelId": {
                        "type": "number"
                    },
                    "color": {
                        "type": "string"
                    },
                    "year": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "required": [
                    "name",
                    "modelId",
                    "color",
                    "year",
                    "image"
                ]
            },
            "BadRequestResponse": {
                "type": "object",
                "properties": {
                    "statusCode": {
                        "type": "number",
                        "default": 400
                    },
                    "message": {
                        "type": "string",
                        "default": "Error description"
                    },
                    "error": {
                        "type": "string",
                        "default": "Bad Request"
                    }
                },
                "required": [
                    "statusCode",
                    "message",
                    "error"
                ]
            },
            "UpdateCarRequest": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "modelId": {
                        "type": "number"
                    },
                    "color": {
                        "type": "string"
                    },
                    "year": {
                        "type": "number"
                    },
                    "id": {
                        "type": "number"
                    },
                    "image": {
                        "type": "string",
                        "format": "binary"
                    }
                },
                "required": [
                    "name",
                    "modelId",
                    "color",
                    "year",
                    "id",
                    "image"
                ]
            }
        }
    }
}